Finished writing unit tests (tests pass) and the Encoder converters and encoders.

+++ mmtf-encoder/src/main/java/org/rcsb/mmtf/encoder/ArrayConverters.java
@@ -64,7 +64,7 @@ public class ArrayConverters {
 		DataOutputStream dos = new DataOutputStream(baos);
 		for(int i=0; i < intArray.length; ++i)
 		{
-			dos.writeShort(intArray[i]);
+			dos.writeInt(intArray[i]);
 		}
 
 		return baos.toByteArray();
